Fix uint64_t -> size-t warnings.

+++ coding/reader.cpp
@@ -6,7 +6,7 @@
 void Reader::ReadAsString(string & s) const
 {
   s.clear();
-  size_t const sz = Size();
+  size_t const sz = static_cast<size_t>(Size());
   s.resize(sz);
   Read(0, &s[0], sz);
 }
